requirements.txt
pymongo==4.6.1
PyYAML==6.0.1
